+++ frontend/app/(tabs)/search.tsx
@@ -6,20 +6,25 @@ import { useEffect, useState } from "react";
 
 
 export default function SearchScreen() {
-
-  const [media, setMedia] = useState<MediaResponse[]>([]);
+  const [songs, setSongs] = useState<MediaResponse[]>([]);
+  const [albums, setAlbums] = useState<MediaResponse[]>([]);
   const BASE_URL = process.env.EXPO_PUBLIC_BASE_URL;
   
   useEffect(() => {
     axios
-      .get(`${BASE_URL}/media?sort=review`)
-      .then((response) => setMedia(response.data))
+      .get(`${BASE_URL}/media?sort=review&type=album`)
+      .then((response) => setAlbums(response.data))
+      .catch((error) => console.error(error));
+
+      axios
+      .get(`${BASE_URL}/media?sort=review&type=track`)
+      .then((response) => setSongs(response.data))
       .catch((error) => console.error(error));
   }, []);
 
   return <View style={styles.container}>
-    <TopSongs songs={media}/>
-    <TopAlbums albums={media}/>
+    <TopSongs songs={songs}/>
+    <TopAlbums albums={albums}/>
     </View>;
 }
